Add AdNauseam to the extension blocklist.
Note: this isn't a security or stability block, but rather a block against unwanted extensions that can and will hurt publishers by devaluating web assets and cutting their income through invalid (faked) traffic.
This is considered unwanted as long as the extension features automated ad clicking. While not click-fraud, since the extension devs do not gain any benefit from it, it is damaging to all publishers that users of the extension visit and therefore considered malware.