Remove certificate issuer organization to common name fallback
Before this change, if a certificate's issuer DN did not have an organization component,
nsIX509Cert.issuerOrganization would fall back to using the issuer common name.
This was never a good idea, because this gave misleading information to consumers of
this interface. Furthermore, it appears that all consumers of this interface already do such
a fallback (for display purposes) when they've determined that it's a reasonable thing to do.