Issue #1872 - Part 1: Update and align about:home and newtab search box styling
Redesigned the styling for the search box for newtab (quickdial) and about:home.
This gets rid of the "play" button and reinstates the "Search" localized term
for submission of the search terms.
Also tweaked the spacing around the logo to neatly tuck into the left side of
the search input bar, which was sloppy and especially visible with image-filling
logotypes (i.e. square).